private user notes

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 // Multiple notes
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+    
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);   // security feature that this is defined
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);    // HTTP 200 "Good"
     } catch (error) {
         next(error);
@@ -16,8 +20,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 // One note
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);   // security feature that this is defined
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");     // HTTP 400 "Bad Request"
@@ -27,6 +33,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");    // HTTP 404 "Not found"
+        }
+
+        if (!note.userId?.equals(authenticatedUserId)) { // check this
+            throw createHttpError(401, "You cannot access this note");
         }
 
         res.status(200).json(note);     // HTTP 200 "Good"
@@ -43,13 +53,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);   // security feature that this is defined
+
         if (!title) {
             throw createHttpError(400, "Note must have a title");     // HTTP 400 "Bad Request"
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title, 
             text: text,
         }); // sends an exec by default, no need to append
@@ -70,11 +84,14 @@ interface UpdateNoteBody {
 }
 
 export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async(req, res, next) => {
+    const noteId = req.params.noteId;
+    const newTitle = req.body.title;
+    const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
+    
     try {
-        const noteId = req.params.noteId;
-        const newTitle = req.body.title;
-        const newText = req.body.text;
-
+        assertIsDefined(authenticatedUserId);
+        
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");     // HTTP 400 "Bad Request"
         }
@@ -87,6 +104,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note not found");    // HTTP 404 "Not found"
+        }
+
+        if (!note.userId?.equals(authenticatedUserId)) { // check this
+            throw createHttpError(401, "You cannot access this note");
         }
         
         note.title = newTitle;
@@ -103,8 +124,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");     // HTTP 400 "Bad Request"
         }
@@ -113,6 +137,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");      // HTTP 404 "Not found"
+        }
+
+        if (!note.userId?.equals(authenticatedUserId)) { // check this
+            throw createHttpError(401, "You cannot access this note");
         }
 
         await NoteModel.findByIdAndDelete(noteId).exec();
